LF-4158 move and rename folders
