+++ packages/webapp/src/components/Animals/AddAnimalsDetails/Origin.tsx
@@ -14,10 +14,10 @@
  */
 
 import { useMemo } from 'react';
-import Input, { getInputErrors } from '../../../Form/Input';
-import RadioGroup from '../../../Form/RadioGroup';
+import Input, { getInputErrors } from '../../Form/Input';
+import RadioGroup from '../../Form/RadioGroup';
 import { DetailsFields, type Option, type CommonDetailsProps } from './type';
-import { AnimalOrigins } from '../../../../containers/Animals/types';
+import { AnimalOrigins } from '../../../containers/Animals/types';
 import styles from './styles.module.scss';
 
 export type OriginProps = CommonDetailsProps & {
